Move iOS GestureNavigationDecoration impl to commonMain
This allows devs to easily test out the implementation from
non-Apple platforms. The class has been renamed to
CupertinoGestureNavigationDecoration to avoid any potential
naming issues.

The GestureNavigationDecoration() function
continues to provide an appropriate impl for each platform.
